editor: add one-key tile rotation
By default in GameMaker, tiles are usually transformed with tile_set_* functions that adds a transformation bitmask to the tile. This effectively makes a regular 360 rotation impossible without chaining the transform commands (rotate, flip, mirror.) After two nights of getting this to work, it's alive, baby.

There are also some miscallenous changes to the default Button object that lets it take in an argument.